refactor: move errors.py to utils/errors.py
