+++ affine/utils/errors.py
